requirements.txt
astroid==2.3.2
cachetools==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.4
django-import-export==2.4.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
httplib2==0.15.0
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==6.2.0
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pylint==2.4.3
pytz==2019.1
PyYAML==5.3.1
rsa==4.0
six==1.12.0
sqlparse==0.3.0
tablib==3.0.0
typed-ast==1.4.0
uritemplate==3.0.1
wrapt==1.11.2
xlrd==2.0.1
xlwt==1.3.0
